@@ -32,8 +32,8 @@
   "homepage": "",
   "dependencies": {
     "@dcl/feature-flags": "^1.2.0",
-    "@dcl/schemas": "^18.1.0",
-    "@dcl/ui-env": "^1.4.0",
+    "@dcl/schemas": "^19.8.0",
+    "@dcl/ui-env": "^1.5.1",
     "@hookform/resolvers": "^3.4.2",
     "@pushprotocol/restapi": "^1.7.3",
     "@snapshot-labs/snapshot.js": "0.10.1",
@@ -49,29 +49,29 @@
     "dayjs-precise-range": "^1.0.1",
     "dcl-catalyst-client": "^21.7.0",
     "decentraland-crypto-middleware": "^1.2.0",
-    "decentraland-dapps": "^25.11.0",
-    "decentraland-ui": "^6.22.0",
-    "decentraland-ui2": "^0.44.1",
+    "decentraland-dapps": "^26.2.2",
+    "decentraland-ui": "^7.1.0",
+    "decentraland-ui2": "^1.1.1",
     "dompurify": "^2.3.3",
     "dotenv": "^16.0.0",
-    "ethers": "^5.7.1",
+    "ethers": "^5.7.2",
     "flickity": "^2.3.0",
     "isbot": "^5.1.22",
     "isomorphic-fetch": "^3.0.0",
     "lodash": "^4.17.21",
     "numeral": "^2.0.6",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-ace": "^10.1.0",
     "react-chartjs-2": "^5.2.0",
-    "react-dom": "^17.0.2",
-    "react-flickity-component": "^3.6.2",
+    "react-dom": "^18.3.1",
+    "react-flickity-component": "^4.0.7",
     "react-helmet": "^6.1.0",
     "react-hook-form": "^7.43.5",
-    "react-intl": "^6.4.4",
+    "react-intl": "^6.5.0",
     "react-loading-skeleton": "^3.1.0",
     "react-markdown": "^7.1.2",
     "react-papaparse": "^4.4.0",
-    "react-redux": "^7.2.9",
+    "react-redux": "^8.1.3",
     "react-responsive": "^9.0.2",
     "react-router-dom": "^6.14.0",
     "react-select": "^5.3.2",
@@ -107,9 +107,9 @@
     "@types/jest": "^29.5.8",
     "@types/node": "^18.8.0",
     "@types/numeral": "^2.0.1",
-    "@types/react": "^18.0.37",
+    "@types/react": "^18.3.12",
     "@types/react-chartjs-2": "^2.5.7",
-    "@types/react-dom": "^18.0.11",
+    "@types/react-dom": "^18.3.1",
     "@types/react-helmet": "^6.1.11",
     "@types/swiper": "^6.0.0",
     "@types/validator": "^13.6.6",
